+++ changelog.md
@@ -1,6 +1,14 @@
 # Changelog
 This changelog uses [Semantic Versioning 2.0.0](https://semver.org/).
 
+## `6.7.1`
+### Changes:
+ + Added new fields to `IPlayerData`.
+ + Added `ObjectStatusData.processObject(data ObjectData)` method to set class property.
+### Fixes:
+ + Fixed a bug in the `PlayerTracker` component where the `worldPos` property was not updating.
+ + Updated packet ids to X21.0.2.
+
 ## `6.7.0`
 ### Changes:
  + Added `PlayerTracker` plugin component. This can be used as part of other plugins to avoid having to implement the player tracking logic for any plugin which requires it. See the `plugin-components` doc for more info.